ml-service/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.0

# Model Persistence
joblib==1.3.2

# Data Processing
python-dateutil==2.8.2

# API and Validation
marshmallow==3.20.1
python-dotenv==1.0.0

# Optional Deep Learning (comment out if not needed)
# tensorflow==2.15.0
# torch==2.1.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
